Replacing editor by reader
I replaced the python editor by a python version of the app in this repo, the python editor will go into the editor repo as they're related. - Python and C++ versions read the same files, just, the python version feature a much lesser media integration and UI, for the benefit of being able to run on older platforms.